fix kunlun not support size_t. test=develop

+++ paddle/fluid/framework/fleet/heter_ps/heter_comm_kernel.cu
@@ -22,7 +22,7 @@ namespace framework {
 #ifdef PADDLE_WITH_CUDA
 
 template <typename T>
-__global__ void fill_idx(T* idx, size_t len) {
+__global__ void fill_idx(T* idx, uint64_t len) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) {
     idx[i] = i;
@@ -30,7 +30,8 @@ __global__ void fill_idx(T* idx, size_t len) {
 }
 
 // template <typename T>
-// void show_tensor(T* input, size_t len, gpuStream_t stream, std::string name)
+// void show_tensor(T* input, uint64_t len, gpuStream_t stream, std::string
+// name)
 // {
 //  T tmp[len];  // NOLINT
 //  cudaMemcpyAsync(&tmp, input, sizeof(T) * len, cudaMemcpyDeviceToHost,
@@ -44,7 +45,7 @@ __global__ void fill_idx(T* idx, size_t len) {
 //}
 
 template <typename T>
-__global__ void calc_shard_offset(T* idx, T* left, T* right, size_t len) {
+__global__ void calc_shard_offset(T* idx, T* left, T* right, uint64_t len) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len - 1) {
     if (idx[i] != idx[i + 1]) {
@@ -61,7 +62,7 @@ __global__ void calc_shard_offset(T* idx, T* left, T* right, size_t len) {
 }
 
 template <typename KeyType, typename T>
-__global__ void calc_shard_index(KeyType* d_keys, size_t len, T* shard_index,
+__global__ void calc_shard_index(KeyType* d_keys, uint64_t len, T* shard_index,
                                  int total_gpu) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) {
@@ -71,7 +72,7 @@ __global__ void calc_shard_index(KeyType* d_keys, size_t len, T* shard_index,
 
 template <typename KeyType, typename T>
 __global__ void fill_shard_key(KeyType* d_shard_keys, KeyType* d_keys, T* idx,
-                               size_t len) {
+                               uint64_t len) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) {
     d_shard_keys[i] = d_keys[idx[i]];
@@ -81,7 +82,7 @@ __global__ void fill_shard_key(KeyType* d_shard_keys, KeyType* d_keys, T* idx,
 template <typename KeyType, typename GradType, typename T>
 __global__ void fill_shard_grads(KeyType* d_shard_keys, KeyType* d_keys,
                                  GradType* d_shard_grads, GradType* d_grads,
-                                 T* idx, size_t len) {
+                                 T* idx, uint64_t len) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) {
     d_shard_keys[i] = d_keys[idx[i]];
@@ -91,7 +92,7 @@ __global__ void fill_shard_grads(KeyType* d_shard_keys, KeyType* d_keys,
 
 template <typename ValType, typename T>
 __global__ void fill_dvals(ValType* d_shard_vals, ValType* d_vals, T* idx,
-                           size_t len) {
+                           uint64_t len) {
   const size_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) {
     d_vals[idx[i]] = d_shard_vals[i];
@@ -100,21 +101,21 @@ __global__ void fill_dvals(ValType* d_shard_vals, ValType* d_vals, T* idx,
 
 // cuda implemention of  heter_comm_kernel.h
 template <typename T, typename StreamType>
-void HeterCommKernel::fill_idx(T* idx, size_t len, StreamType& stream) {
+void HeterCommKernel::fill_idx(T* idx, uint64_t len, StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
   fill_idx<<<grid_size, block_size_, 0, stream>>>(idx, len);
 }
 
 template <typename T, typename StreamType>
-void HeterCommKernel::calc_shard_offset(T* idx, T* left, T* right, size_t len,
-                                        size_t total_devs, StreamType& stream) {
+void HeterCommKernel::calc_shard_offset(T* idx, T* left, T* right, uint64_t len,
+                                        int total_devs, StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
   calc_shard_offset<<<grid_size, block_size_, 0, stream>>>(idx, left, right,
                                                            len);
 }
 
 template <typename KeyType, typename T, typename StreamType>
-void HeterCommKernel::calc_shard_index(KeyType* d_keys, size_t len,
+void HeterCommKernel::calc_shard_index(KeyType* d_keys, uint64_t len,
                                        T* shard_index, int total_gpu,
                                        StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
@@ -124,7 +125,7 @@ void HeterCommKernel::calc_shard_index(KeyType* d_keys, size_t len,
 
 template <typename KeyType, typename T, typename StreamType>
 void HeterCommKernel::fill_shard_key(KeyType* d_shard_keys, KeyType* d_keys,
-                                     T* idx, size_t len, StreamType& stream) {
+                                     T* idx, uint64_t len, StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
   fill_shard_key<<<grid_size, block_size_, 0, stream>>>(d_shard_keys, d_keys,
                                                         idx, len);
@@ -133,7 +134,7 @@ void HeterCommKernel::fill_shard_key(KeyType* d_shard_keys, KeyType* d_keys,
 template <typename KeyType, typename GradType, typename T, typename StreamType>
 void HeterCommKernel::fill_shard_grads(KeyType* d_shard_keys, KeyType* d_keys,
                                        GradType* d_shard_grads,
-                                       GradType* d_grads, T* idx, size_t len,
+                                       GradType* d_grads, T* idx, uint64_t len,
                                        StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
   fill_shard_grads<<<grid_size, block_size_, 0, stream>>>(
@@ -142,7 +143,7 @@ void HeterCommKernel::fill_shard_grads(KeyType* d_shard_keys, KeyType* d_keys,
 
 template <typename ValType, typename T, typename StreamType>
 void HeterCommKernel::fill_dvals(ValType* d_shard_vals, ValType* d_vals, T* idx,
-                                 size_t len, StreamType& stream) {
+                                 uint64_t len, StreamType& stream) {
   int grid_size = (len - 1) / block_size_ + 1;
   fill_dvals<<<grid_size, block_size_, 0, stream>>>(d_shard_vals, d_vals, idx,
                                                     len);
